+++ src/test/java/com/iexec/worker/chain/IexecHubServiceTests.java
@@ -46,6 +46,8 @@ import java.math.BigInteger;
 import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -73,6 +75,8 @@ class IexecHubServiceTests {
     private Web3j web3jClient;
     private IexecHubService iexecHubService;
     private Credentials credentials;
+    @Mock
+    private CompletableFuture<?> completableFuture;
 
     @BeforeEach
     void init() throws Exception {
@@ -127,6 +131,43 @@ class IexecHubServiceTests {
         IexecHubContract.TaskContributeEventResponse response = iexecHubService.contribute(contribution);
         assertThat(response).isNotNull();
     }
+
+    @Test
+    void shouldNotContributeOnExecutionException() throws ExecutionException, InterruptedException {
+        final Contribution contribution = Contribution.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .enclaveChallenge("enclaveChallenge")
+                .enclaveSignature("enclaveSignature")
+                .resultHash("resultHash")
+                .resultSeal("resultSeal")
+                .workerPoolSignature("workerPoolSignature")
+                .build();
+        try (MockedStatic<CompletableFuture> mockedStatic = mockStatic(CompletableFuture.class)) {
+            mockedStatic.when(() -> CompletableFuture.supplyAsync(any(), any())).thenReturn(completableFuture);
+            when(completableFuture.get()).thenThrow(ExecutionException.class);
+            IexecHubContract.TaskContributeEventResponse response = iexecHubService.contribute(contribution);
+            assertThat(response).isNull();
+        }
+    }
+
+    @Test
+    void shouldNotContributeWhenInterrupted() throws ExecutionException, InterruptedException {
+        final Contribution contribution = Contribution.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .enclaveChallenge("enclaveChallenge")
+                .enclaveSignature("enclaveSignature")
+                .resultHash("resultHash")
+                .resultSeal("resultSeal")
+                .workerPoolSignature("workerPoolSignature")
+                .build();
+        try (MockedStatic<CompletableFuture> mockedStatic = mockStatic(CompletableFuture.class)) {
+            mockedStatic.when(() -> CompletableFuture.supplyAsync(any(), any())).thenReturn(completableFuture);
+            when(completableFuture.get()).thenThrow(InterruptedException.class);
+            IexecHubContract.TaskContributeEventResponse response = iexecHubService.contribute(contribution);
+            assertThat(response).isNull();
+            assertThat(Thread.currentThread().isInterrupted()).isTrue();
+        }
+    }
     // endregion
 
     // region reveal
@@ -144,6 +185,27 @@ class IexecHubServiceTests {
 
         IexecHubContract.TaskRevealEventResponse response = iexecHubService.reveal(CHAIN_TASK_ID, "resultDigest");
         assertThat(response).isNotNull();
+    }
+
+    @Test
+    void shouldNotRevealOnExecutionException() throws ExecutionException, InterruptedException {
+        try (MockedStatic<CompletableFuture> mockedStatic = mockStatic(CompletableFuture.class)) {
+            mockedStatic.when(() -> CompletableFuture.supplyAsync(any(), any())).thenReturn(completableFuture);
+            when(completableFuture.get()).thenThrow(ExecutionException.class);
+            IexecHubContract.TaskRevealEventResponse response = iexecHubService.reveal(CHAIN_TASK_ID, "resultDigest");
+            assertThat(response).isNull();
+        }
+    }
+
+    @Test
+    void shouldNotContributeRevealWhenInterrupted() throws ExecutionException, InterruptedException {
+        try (MockedStatic<CompletableFuture> mockedStatic = mockStatic(CompletableFuture.class)) {
+            mockedStatic.when(() -> CompletableFuture.supplyAsync(any(), any())).thenReturn(completableFuture);
+            when(completableFuture.get()).thenThrow(InterruptedException.class);
+            IexecHubContract.TaskRevealEventResponse response = iexecHubService.reveal(CHAIN_TASK_ID, "resultDigest");
+            assertThat(response).isNull();
+            assertThat(Thread.currentThread().isInterrupted()).isTrue();
+        }
     }
     // end region
 
@@ -167,6 +229,43 @@ class IexecHubServiceTests {
                 .build();
         Optional<ChainReceipt> chainReceipt = iexecHubService.contributeAndFinalize(contribution, "resultLink", "callbackData");
         assertThat(chainReceipt).isNotEmpty();
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeOnExecutionException() throws ExecutionException, InterruptedException {
+        final Contribution contribution = Contribution.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .enclaveChallenge("enclaveChallenge")
+                .enclaveSignature("enclaveSignature")
+                .resultHash("resultHash")
+                .resultSeal("resultSeal")
+                .workerPoolSignature("workerPoolSignature")
+                .build();
+        try (MockedStatic<CompletableFuture> mockedStatic = mockStatic(CompletableFuture.class)) {
+            mockedStatic.when(() -> CompletableFuture.supplyAsync(any(), any())).thenReturn(completableFuture);
+            when(completableFuture.get()).thenThrow(ExecutionException.class);
+            Optional<ChainReceipt> chainReceipt = iexecHubService.contributeAndFinalize(contribution, "resultLink", "callbackData");
+            assertThat(chainReceipt).isEmpty();
+        }
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeWhenInterrupted() throws ExecutionException, InterruptedException {
+        final Contribution contribution = Contribution.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .enclaveChallenge("enclaveChallenge")
+                .enclaveSignature("enclaveSignature")
+                .resultHash("resultHash")
+                .resultSeal("resultSeal")
+                .workerPoolSignature("workerPoolSignature")
+                .build();
+        try (MockedStatic<CompletableFuture> mockedStatic = mockStatic(CompletableFuture.class)) {
+            mockedStatic.when(() -> CompletableFuture.supplyAsync(any(), any())).thenReturn(completableFuture);
+            when(completableFuture.get()).thenThrow(InterruptedException.class);
+            Optional<ChainReceipt> chainReceipt = iexecHubService.contributeAndFinalize(contribution, "resultLink", "callbackData");
+            assertThat(chainReceipt).isEmpty();
+            assertThat(Thread.currentThread().isInterrupted()).isTrue();
+        }
     }
     // endregion