+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [[8.1.2]](https://github.com/iExecBlockchainComputing/iexec-worker/releases/tag/v8.1.2) 2023-06-22
+
+### Features
+- Retry dataset download on several IPFS gateways. (#532)
+### Bug Fixes
+- Improve interruptions handling in `IexecHubService`. (#529)
+- Return application execution `exitCause` as computed in `AppComputeService#getExitCauseFromFinalStatus`. (#531)
+### Dependency upgrade
+- Upgrade to `iexec-commons-poco` 3.0.3 dependency. (#532)
+
 ## [[8.1.1]](https://github.com/iExecBlockchainComputing/iexec-worker/releases/tag/v8.1.1) 2023-06-15
 
 ### Bug Fixes